# internal/handler/v1/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger specification for a Duel Duck back-end.",
        "title": "Duel Duck API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/coins": {
            "post": {
                "description": "Retrieve a list of all coins with support for filtering, ordering, and pagination via query parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Get all coins with optional filtering, ordering, and pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "opts.pagination.page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (starting from 1)",
                        "name": "opts.pagination.page_num",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field to order by",
                        "name": "opts.order.order_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "desc",
                            "asc"
                        ],
                        "type": "string",
                        "default": "\"\"",
                        "description": "Order type",
                        "name": "opts.order.order_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter column name",
                        "name": "opts.filters[0].column",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter operator",
                        "name": "opts.filters[0].operator",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter value",
                        "name": "opts.filters[0].value",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter OR condition",
                        "name": "opts.filters[0].where_or",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of coins retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Coin"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request params",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/coins/mint/{mint}": {
            "get": {
                "description": "Retrieve Solana token information by searching for tokens with the specified mint address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Get Solana token by mint address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mint address to search for",
                        "name": "mint",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token information retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/model.SolanaToken"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or invalid mint address format",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "404": {
                        "description": "Token with provided mint address not found",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error during token retrieval",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/coins/solana-price": {
            "get": {
                "description": "Retrieve the current price of Solana cryptocurrency",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Get current Solana price",
                "responses": {
                    "200": {
                        "description": "Current Solana price retrieved successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "solana_price": {
                                    "type": "number"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error during price retrieval",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "503": {
                        "description": "External service unavailable",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/coins/{name}": {
            "get": {
                "description": "Retrieve Solana token information by searching for tokens with the specified name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Get Solana token by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token name to search for",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tokens matching the name",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SolanaToken"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data or empty token name",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error during token retrieval",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperrors.ErrorPublic": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.Coin": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
...
        "model.SolanaToken": {
            "type": "object",
            "properties": {
                "daily_volume": {
                    "type": "number"
                },
                "decimals": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "mint": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
...
    }
}